src/modules/product-adm/facade/product-adm.facade.ts
import UseCaseInterface from "../../@shared/domain/usecase/usecase.interface";
import ProductAdmFacadeInterface, {
  AddProductFacadeInputDto,
  CheckStockFacadeInputDto,
  CheckStockFacadeOutputDto,
} from "./product-adm.facade.interface";

export interface UseCasesProps {
  addUseCase: UseCaseInterface;
  stockUseCase: UseCaseInterface;
}

export default class ProductAdmFacade implements ProductAdmFacadeInterface {
  private _addUsecase: UseCaseInterface;
  private _checkStockUsecase: UseCaseInterface;

  constructor(usecasesProps: UseCasesProps) {
    this._addUsecase = usecasesProps.addUseCase;
    this._checkStockUsecase = usecasesProps.stockUseCase;
  }

  addProduct(input: AddProductFacadeInputDto): Promise<void> {
    // caso o DTO do usecase for !== do DTO da facade, é preciso converter o DTO da facade para o DTO do usecase
    return this._addUsecase.execute(input);
  }

  checkStock(
    input: CheckStockFacadeInputDto
  ): Promise<CheckStockFacadeOutputDto> {
    // caso o DTO do usecase for !== do DTO da facade, é preciso converter o DTO da facade para o DTO do usecase
    return this._checkStockUsecase.execute(input);
  }
}
